requirements.txt
flask
python-dotenv
pinecone-client
langchain-community
ctransformers
langchain_huggingface
sentence-transformers
